\name{atom_mw_table}
\alias{atom_mw_table}
\docType{data}
\title{
Table of atomic weights
}
\description{
This data.frame contains 42 atoms with their monoisotpic weights. These atoms
are commonly found in proteins.
}
\usage{data("atom_mw_table")}
\format{
    A data frame with 42 observations on the following 2 variables.
    \describe{
        \item{\code{Element}}{a character vector}
        \item{\code{Monoisotopic}}{a numeric vector}
    }
}

\examples{
data(atom_mw_table)
}
\keyword{datasets}
