% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mms2plot.R
\name{mms2plot}
\alias{mms2plot}
\title{mms2plot}
\usage{
mms2plot(id_table_path, par_xml_path, mqpar_filepath, output_path,
  min_intensity_ratio = 0.01, pdf_width = 3.35,
  pdf_height = pdf_width/2.4, xmai = 0.15 * pdf_width/3.35,
  ymai = 0.3 * pdf_width/3.35, ppm = 20, y_ion_col = "red",
  b_ion_col = "blue", peaks_col = "grey", ymax = 1.6,
  info_height = 1.5, peptide_height = 1.3, mod_height = 0.07,
  len_annoSpace = 0.1, lwd = 1 * pdf_width/3.35, cex = 1 *
  pdf_width/3.35, show_letterBY = FALSE)
}
\arguments{
\item{id_table_path}{File path name of a table that contains MS2 information
of identified (un)modified peptides plus a group labelling. The format
of MS2 information is referred to as the output file ms2.txt from
the Maxquant search software.}

\item{par_xml_path}{Xml file path of parameters for modifications and
labelling. The file is the Maxquant parameter file modifications.xml.}

\item{mqpar_filepath}{File path name that includes a list of parameter files
for search engines. The parameter file format is referred to as
mqpar.xml in Maxquant.}

\item{output_path}{path of output files.}

\item{min_intensity_ratio}{minimal percentage threshold of MS2 intensity,
compared with the highest intensity. (default=0.01).}

\item{pdf_width}{The width of a single PSM figure area in inches. The area
includes both plot region and outer margin area. The default is 3.35
for a single column. The width is 7 for double-column.}

\item{pdf_height}{The height of a single PSM figure area in inches. The
default is pdf_width/2.4.}

\item{xmai}{margin of the figure in number of inches for x axis.
(default=pdf_width*0.15/3.35).}

\item{ymai}{margin of the figure in number of inches for y axis.
(default=pdf_width*0.15/3.35).}

\item{ppm}{The threshold of mass error in parts per million(ppm):
(exactMass-accurateMass)/exactMass*1E6. (default=20).}

\item{y_ion_col}{y ion color. The default is red.}

\item{b_ion_col}{b ion color. The default is blue.}

\item{peaks_col}{color of all M/Z peaks. The default is grey.}

\item{ymax}{The height of the plot region relative to the default.
(default=1.6).}

\item{info_height}{The height of MS2 annotation. (default=1.5)}

\item{peptide_height}{The height of peptide sequence annotation in the plot
region, relative to the default. (default=1.3)}

\item{mod_height}{The height of modification annotation relative to the
location where peptide sequence is annotation. (default=0.07).}

\item{len_annoSpace}{The length of b/y ion annotation segments.(default=0.1).}

\item{lwd}{line width relative to the default. (default=pdf_width/3.35).}

\item{cex}{A numerical value giving the amount by which plotting text and
symbols is magnified relative to the default.(default=pdf_width/3.35).}

\item{show_letterBY}{Logical: should "b" or "y" characters are shown on the
peak annotation? The default is FALSE.}
}
\value{
No value is returned.
}
\description{
Visualization of multiple MS/MSs for (un)modified peptides
}
\note{
See vignettes for more details
}
\examples{
general_path = system.file( package = "mms2plot",dir = "extdata" )
###################################
# Generate mms2plot for label-free data
lf_path = system.file( package = "mms2plot",dir = "extdata/label_free" )
# id_table_path expands the Maxqaunt output msms.txt by adding "label" column
id_table_path = dir( lf_path, "msms_TMT.txt", full.names = TRUE )
par_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
# mqpar_filepath contains mqpar.xml with full file path and PPM cutoff
mqpar_filepath = dir( general_path, "mqpar_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot( id_table_path, par_xml_path, mqpar_filepath, output_path) #not run

###################################
# Generate mms2plot for TMT labelling
TMT_path = system.file( package = "mms2plot",dir = "extdata/TMT" )
# id_table_path expands the Maxqaunt output msms.txt by adding "label" column
id_table_path = dir( TMT_path, "msms_TMT.txt", full.names = TRUE )
par_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
# mqpar_filepath contains mqpar.xml with full file path and PPM cutoff
mqpar_filepath = dir( general_path, "mqpar_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot( id_table_path, par_xml_path, mqpar_filepath, output_path) #not run

#####################################
# Generate mms2plot for SILAC labelling
SILAC_path = system.file( package = "mms2plot",dir = "extdata/silac" )
id_table_path = dir( SILAC_path, "msms_SILAC.txt", full.names = TRUE )
par_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
mqpar_filepath = dir( general_path, "mqpar_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot( id_table_path, par_xml_path,mqpar_filepath,output_path ) #not run

#####################################
# Generate mms2plot for dimethyl labelling
dim_path = system.file(package="mms2plot",dir="extdata/Dimethyl_Labelling")
id_table_path = dir( dim_path, "msms_dim.txt", full.names = TRUE )
par_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
mqpar_filepath = dir( general_path, "mqpar_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot(id_table_path, par_xml_path, mqpar_filepath, output_path) #not run

}
\author{
Lei Li
}
