% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZID_prep.R
\name{MZID_prep}
\alias{MZID_prep}
\title{Conversion of the research result with the mzid format to the mms2plot-required identification file.}
\usage{
MZID_prep(path_user_table, output_path)
}
\arguments{
\item{output_path}{a character string naming a output file as the identification txt file for mms2plot.}

\item{config_table_path}{file path of the table that contains five columns: the raw MS filepath, 
the type of modifiation, mass shift, the tolerance of the mass shift and the search result MZID filepath. 
Please see teh user guide for the details.}
}
\value{
No value is returned.
}
\description{
Conversion of the research result with the mzid format to the mms2plot-required identification file.
}
\examples{
general_path = system.file( package = "mms2plot",dir = "extdata" )
setwd( general_path )
config_table_path = 'prep/user_table_forMSMS.txt'
path_user_table = 'prep/user_table_forMZID.txt'
output_path = 'prep/comet/conversion/identification.txt'
MZID_prep(path_user_table,output_path)#,dir_mzid)#label.by.Raw_seq_charge)



}
