\name{aa_mw_table}
\alias{aa_mw_table}
\docType{data}
\title{
aa_mw_table
}
\description{
amino acid molecular weight table
}
\usage{data("aa_mw_table")}
\format{
  A data frame with 25 observations on the following 7 variables.
  \describe{
    \item{\code{aa}}{a character vector}
    \item{\code{weight}}{a numeric vector}
    \item{\code{aa_varmod}}{a character vector}
    \item{\code{labelmod}}{a logical vector}
    \item{\code{reporterion}}{a logical vector}
    \item{\code{reporterion_group}}{a logical vector}
    \item{\code{labelmod_group}}{a logical vector}
  }
}


\examples{
data(aa_mw_table)

}
\keyword{datasets}
