% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mms2plot.R
\docType{data}
\name{aa_mw_table}
\alias{aa_mw_table}
\title{mms2plot}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 25 rows and 7 columns.}
\usage{
aa_mw_table
}
\arguments{
\item{aa_mw_table}{Molecular weight of Amino acids}

\item{atom_mw_table}{Molecular weight of atoms}

\item{id_table_path}{File path name of a table that contains MS2 information
of identified (un)modified peptides plus a group labelling. The format
of MS2 information is referred to as the output file ms2.txt from
the Maxquant search software.}

\item{par_xml_path}{Xml file path of parameters for modifications and
labelling. The file format is referred to as modifications.xml in
Maxquant.}

\item{mqpar_filepath}{File path name that includes a list of parameter files
for search engines. The parameter file format is referred to as
mqpar.xml in Maxquant.}

\item{min_intensity_ratio}{minimal percentage threshold of MS2 intensity,
compared with the highest intensity. (default=0.01).}

\item{pdf_width}{The width of a single PSM figure area in inches. The area
includes both plot region and outer margin area. The default is 3.35
for a single column. The width is 7 for double-column.}

\item{pdf_height}{The height of a single PSM figure area in inches. The
default is pdf_width/2.4.}

\item{xmai}{margin of the figure in number of inches for x axis.
(default=pdf_width*0.15/3.35).}

\item{ymai}{margin of the figure in number of inches for y axis.
(default=pdf_width*0.15/3.35).}

\item{ppm}{The threshold of mass error in parts per million(ppm):
(exactMass-accurateMass)/exactMass*1E6. (default=20).}

\item{y_ion_col}{y ion color. The default is red.}

\item{b_ion_col}{b ion color. The default is blue.}

\item{peaks_col}{color of all M/Z peaks. The default is grey.}

\item{ymax}{The height of the plot region relative to the default.
(default=1.6).}

\item{info_height}{The height of MS2 annotation. (default=1.5)}

\item{peptide_height}{The height of peptide sequence annotation in the plot
region, relative to the default. (default=1.3)}

\item{mod_height}{The height of modification annotation relative to the
location where peptide sequence is annotation. (default=0.07).}

\item{len_annoSpace}{The length of b/y ion annotation segments. (default=0.1).}

\item{lwd}{line width relative to the default. (default=pdf_width/3.35).}

\item{cex}{A numerical value giving the amount by which plotting text and
symbols is magnified relative to the default. (default=pdf_width/3.35).}

\item{show_letterBY}{Logical: should "b"/"y" characters are shown on the peak
annotation? The default is FALSE.}
}
\description{
Visualization of multiple MS/MSs for (un)modified peptides
}
\author{
Lei Li
}
\keyword{datasets}
