% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mms2plot.R
\name{mms2plot}
\alias{mms2plot}
\title{mms2plot}
\usage{
mms2plot(
  id_table_path,
  mod_xml_path,
  par_filepath,
  output_path,
  min_intensity_ratio = 0.01,
  pdf_width = 3.35,
  pdf_height = pdf_width/2.4,
  xmai = 0.15 * pdf_width/3.35,
  ymai = 0.3 * pdf_width/3.35,
  ppm = 20,
  y_ion_col = "red",
  b_ion_col = "blue",
  peaks_col = "grey",
  ymax = 1.6,
  info_height = 1.5,
  peptide_height = 1.3,
  mod_height = 0.07,
  len_annoSpace = 0.1,
  lwd = 1 * pdf_width/3.35,
  cex = 1 * pdf_width/3.35
)
}
\arguments{
\item{id_table_path}{File path of the table that contains the information of 
peptide-spectrum matches (PSMs) and number labels. The PSM information 
is referred to as the output file msms.txt from Maxquant.}

\item{mod_xml_path}{File path of the modification file. The Maxquant 
mdification xml file modifications.xml can be directly used.}

\item{par_filepath}{File path of of the parameter batch table that includes
the parameter xml file and the fragment mass tolerance (ppm). The 
parameter file format is referred to as par.xml in Maxquant.}

\item{output_path}{Folder Path that stores the output image files.}

\item{min_intensity_ratio}{minimal percentage threshold of MS2 intensity,
compared with the highest intensity. (default=0.01).}

\item{pdf_width}{The width of a single PSM figure area in inches. The area
includes both plot region and outer margin area. The default is 3.35
for a single column. The width is 7 for double-column.}

\item{pdf_height}{The height of a single PSM figure area in inches. The
default is pdf_width/2.4.}

\item{xmai}{Margin of the figure in number of inches for x axis.
(default=pdf_width*0.15/3.35).}

\item{ymai}{Margin of the figure in number of inches for y axis.
(default=pdf_width*0.15/3.35).}

\item{ppm}{The threshold of mass error in parts per million(ppm):
(exactMass-accurateMass)/exactMass*1E6. (default=20).}

\item{y_ion_col}{y ion color. The default is red.}

\item{b_ion_col}{b ion color. The default is blue.}

\item{peaks_col}{color of all M/Z peaks. The default is grey.}

\item{ymax}{The height of the plot region relative to the default.
(default=1.6).}

\item{info_height}{The height of MS2 annotation. (default=1.5)}

\item{peptide_height}{The height of peptide sequence annotation in the plot
region, relative to the default. (default=1.3)}

\item{mod_height}{The height of modification annotation relative to the
location where peptide sequence is annotation. (default=0.07).}

\item{len_annoSpace}{The length of b/y ion annotation segments.(default=0.1).}

\item{lwd}{line width relative to the default. (default=pdf_width/3.35).}

\item{cex}{A numerical value giving the amount by which plotting text and
symbols is magnified relative to the default.(default=pdf_width/3.35).}
}
\value{
No value is returned.
}
\description{
Visualization of multiple MS/MS spectra for groups of modified 
and non-modified peptides
}
\note{
See vignettes for more details
}
\examples{
###
general_path = system.file( package = "mms2plot",dir = "extdata" )
setwd( general_path )
###################################
# Generate a mirrored-spectra pdf file for label-free data
lf_path = system.file( package = "mms2plot",dir = "extdata/label_free" )
# id_table_path expands the Maxqaunt output msms.txt by adding "label" column
id_table_path = dir( lf_path, "msms_labelfree.txt", full.names = TRUE )
mod_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
# par_filepath contains par.xml with full file path and PPM cutoff
par_filepath = dir( general_path, "par_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot( id_table_path, mod_xml_path, par_filepath, output_path) # Only run for testing

###################################
# Generate an aligned-spectra plot pdf file for TMT labelling
TMT_path = system.file( package = "mms2plot",dir = "extdata/TMT" )
# id_table_path expands the Maxqaunt output msms.txt by adding "label" column
id_table_path = dir( TMT_path, "msms_TMT.txt", full.names = TRUE )
mod_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
# par_filepath contains par.xml with full file path and PPM cutoff
par_filepath = dir( general_path, "par_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot( id_table_path, mod_xml_path, par_filepath, output_path) # Only run for testing

#####################################
# Generate a mirrored-spectra pdf file for SILAC labelling
SILAC_path = system.file( package = "mms2plot",dir = "extdata/silac" )
id_table_path = dir( SILAC_path, "msms_SILAC.txt", full.names = TRUE )
mod_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
par_filepath = dir( general_path, "par_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot( id_table_path, mod_xml_path,par_filepath,output_path ) # Only run for testing

#####################################
# Generate a couple of spectra pdf files for dimethyl labelling
dim_path = system.file(package="mms2plot",dir="extdata/Dimethyl_Labelling")
id_table_path = dir( dim_path, "msms_dim.txt", full.names = TRUE )
mod_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
par_filepath = dir( general_path, "par_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot(id_table_path, mod_xml_path, par_filepath, output_path) # Only run for testing

#####################################
# Generate a couple of spectra pdf files for ETD (ie. c/z ions)
dim_path = system.file(package="mms2plot",dir="extdata/label_free_ETD")
id_table_path = dir( dim_path, "msms_lf_ETD.txt", full.names = TRUE )
mod_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
par_filepath = dir( general_path, "par_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot(id_table_path, mod_xml_path, par_filepath, output_path) # Only run for testing

#####################################
# an example for showing neutral loss (e.g. phosphorylation)
dim_path = system.file(package="mms2plot",dir="extdata/phospho")
id_table_path = dir( dim_path, "msms_phospho.txt", full.names = TRUE )
mod_xml_path = dir( general_path, "modifications.xml", full.names = TRUE )
par_filepath = dir( general_path, "par_batch.txt", full.names = TRUE )
output_path = general_path
#mms2plot(id_table_path, mod_xml_path, par_filepath, output_path) # Only run for testing

}
\author{
Lei Li
}
