% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_prep.R
\name{MQ_prep}
\alias{MQ_prep}
\title{Conversion of Maxquant search result file (i.e. msms.txt) to the mms2plot-required identification file.}
\usage{
MQ_prep(config_table_path, msms_path, output_file)
}
\arguments{
\item{config_table_path}{File path of the table that contains the raw MS filepath, 
the type of modifiation, mass shift and the tolerance of the mass shift. 
Please see teh user guide for the details.}

\item{msms_path}{File path of the MaxQuant search result file (i.e. msms.txt).}

\item{output_file}{File path of of the parameter batch table that includes
the parameter xml file and the fragment mass tolerance (ppm). The 
parameter file format is referred to as par.xml in Maxquant.}

\item{output_path}{a character string naming a output file.}
}
\value{
No value is returned.
}
\description{
Conversion of Maxquant search result file (i.e. msms.txt) to the mms2plot-required identification file.
}
\examples{
general_path = system.file( package = "mms2plot",dir = "extdata" )
setwd( general_path )
config_table_path = 'prep/user_table_forMSMS.txt'
msms_path = 'prep/MaxQuant/msms.txt'
output_file = 'prep/MaxQuant/conversion/identification.txt'
MQ_prep(config_table_path,msms_path,output_file)
\dontrun{
general_path = system.file( package = "mms2plot",dir = "extdata" )
setwd( general_path )
config_table_path = 'prep/user_table_forMSMS.txt'
msms_path = 'prep/MaxQuant/msms.txt'
output_file = 'prep/MaxQuant/conversion/identification.txt'
MQ_prep(config_table_path,msms_path,output_file)
}

}
